new password and delete user

+++ src/routes/(static)/settings/+page.svelte
@@ -1,10 +1,16 @@
 <script lang="ts">
 
-	import { signOut, updateEmail, deleteUser } from "firebase/auth";
+	import { signOut, updateEmail, updatePassword, deleteUser } from "firebase/auth";
 	import { auth, user } from "$firebase";
 	
 	let _user: any = $user;
 	let email = _user.email;
+
+	let deleteUserPassword: string;
+	let newEmail: string;
+	let newPasswordCurrentPassword: string;
+	let newPassword: string;
+	let confirmNewPassword: string;
 	
 	user.subscribe((data) => {
 		_user = data;
@@ -21,22 +27,35 @@
 	}
 	
 	const changeEmail = () => {
-    const newEmail = document.getElementById('new-email-input').value;
-	updateEmail(_user, newEmail).then(() => {
-		console.log('Email updated successfully!');
-		email = _user.email;
-	}).catch((error) => {
-		console.log('Error updating email:', error);
-	});
+		updateEmail(_user, newEmail).then(() => {
+			console.log('Email updated successfully!');
+			email = _user.email;
+		}).catch((error) => {
+			console.log('Error updating email:', error);
+		});
+  }
+
+  const changePassword = () => {
+	if (newPasswordCurrentPassword == _user.password && newPassword == confirmNewPassword) {
+		updatePassword(_user, newPassword).then(() => {
+			console.log('Email updated successfully!');
+			email = _user.email;
+		}).catch((error) => {
+			console.log('Error updating email:', error);
+		});
+	}
   }
 
 	const userdel = () => {  
-	deleteUser(_user).then(() => {
-		// User deleted.
-	}).catch((error) => {
-		// An error ocurred
-		// ...
-	});
+	
+	if (_user.password == deleteUserPassword) {
+		deleteUser(_user).then(() => {
+			// User deleted.
+		}).catch((error) => {
+			// An error ocurred
+			// ...
+		});
+	}
 }
 
   </script>
@@ -66,7 +85,7 @@
 					<div class="flex self-center form-control w-full gap-4">
 						<span class="font-bold text-md">New Email:</span>
 						<div class="flex flex-row justify-between">
-							<input type="text" placeholder="email..." class="input input-bordered" id="new-email-input"/>
+							<input type="text" placeholder="email..." class="input input-bordered" bind:value={newEmail}/>
 							<label for="edit-email-popup" class="btn btn-primary self-center" on:click={changeEmail}>Update Email</label>
 						</div>
 					</div>
@@ -92,11 +111,12 @@
 			<div class="modal">
 				<div class="modal-box relative">
 					<label for="edit-password-popup" class="btn btn-sm btn-circle absolute right-2 top-2">✕</label>
+
 					<!-- Popup Content -->
 					<div class="flex self-center form-control w-full gap-4">
-						<span class="font-bold text-md">New Password:</span>
+						<span class="font-bold text-md">Current Password:</span>
 						<div class="flex flex-row">
-							<input type="password" placeholder="password..." class="input input-bordered" id="first-pass" />
+							<input type="password" placeholder="password..." class="input input-bordered" bind:value={newPasswordCurrentPassword} />
 							
 							
 							<label class="btn btn-circle btn-ghost swap swap-rotate absolute left-48" >
@@ -113,10 +133,30 @@
 								</label>
 							
 						</div>
+					
+						<span class="font-bold text-md">New Password:</span>
+						<div class="flex flex-row">
+							<input type="password" placeholder="password..." class="input input-bordered" bind:value={newPassword}/>
+							
+							
+							<label class="btn btn-circle btn-ghost swap swap-rotate absolute left-48" >
+
+								<!-- this hidden checkbox controls the state -->
+								<input type="checkbox" />
+								
+								<!--  icon -->
+								<div class="swap-on fill-current"></div>
+								
+								<!--  icon -->
+								<div class="swap-off fill-current"></div>
+								
+							</label>
+							
+						</div>
 						<span class="font-bold text-md">Confirm Password:</span>
 						<div class="flex flex-row justify-between">
-							<input type="password" placeholder="password..." class="input input-bordered" id="new-pass-input" />
-							<label for="edit-password-popup" class="btn btn-primary self-center">Update Password</label>
+							<input type="password" placeholder="password..." class="input input-bordered" bind:value={confirmNewPassword} />
+							<label for="edit-password-popup" class="btn btn-primary self-center" on:click={changePassword}>Update Password</label>
 						</div>
 					</div>
 				</div>
@@ -124,14 +164,31 @@
 		</div>
 	</div>
 
-	<div class="flex flex-col gap-2">
-		<btn class="flex justify-between btn btn-ghost normal-case p-2 border-red-500/10" on:click={Logout}>
-			<span class="w-fit opacity-70 text-center">Logout</span>
-		</btn>
+	<div class="flex flex-col gap-6">
+		
+		<!-- Logout -->
+		<div class="flex justify-between">
+			<label class="btn btn-ghost normal-case p-2 border-red-500/10 w-full opacity-70 text-center" on:click={Logout}>Logout</label>
+		</div>
 
-		<btn class="flex justify-between btn btn-ghost p-2 border-red-500" on:click={userdel}>
-			<button><span class="w-fit opacity-70 text-center">Delete Account</span></button>
-		</btn>
+		<div class="flex justify-between flex-col">
+			<label for="delete-account-popup" class="w-full opacity-70 text-center btn btn-ghost p-2 border-red-500">Delete Account</label>
+			
+			<input type="checkbox" id="delete-account-popup" class="modal-toggle" />
+			<div class="modal">
+				<div class="modal-box relative">
+					<label for="delete-account-popup" class="btn btn-sm btn-circle absolute right-2 top-2">✕</label>
+					<!-- Popup Content -->
+					<div class="flex self-center form-control w-full gap-4">
+						<span class="font-bold text-md">Enter Password:</span>
+						<div class="flex flex-row justify-between">
+							<input type="password" placeholder="password..." class="input input-bordered" bind:value={deleteUserPassword}/>
+							<label for="delete-account-popup" class="btn btn-primary self-center" on:click={userdel}>Delete Account</label>
+						</div>
+					</div>
+				</div>
+			</div>
+		</div>
 	</div>
 
 </div>
